alpha-sentinel-os/requirements.txt
pandas
langgraph
langchain
langchain-openai
langchain-core
langchain-community
openai
psycopg2-binary
python-dotenv
yfinance 
sqlalchemy 
fredapi
praw
flask
requests
typing-extensions